Create 03. Jupyter Notebooks ♡ SQL (part i)
We've seen that notebooks can display basic objects such as numbers and strings. But notebooks also support and display the outputs of SQL commands! Using an open source Jupyter extension called ipython-sql, we can connect to a database and issue SQL commands within our notebook. For example, we can connect to a PostgreSQL database that has a table that contains country data, then inspect the first three rows of the table by putting %%sql ahead of the SQL commands (more on the meaning of %% later).
%%sql postgresql:///countries
SELECT * FROM countries LIMIT 3;
%%sql postgresql:///countries