Create 07. Jupyter Notebooks ♡ plots
Tables are nice but — as the saying goes — "a plot can show a thousand data points." Notebooks handle plots as well, but it requires some more magic. Here magic does not refer to any arcane rituals but to so-called "magic commands" that affect how the Jupyter Notebook works. Magic commands start with either % or %% (just like we saw with %sql and %%sql) and the command we need to nicely display plots inline is %matplotlib inline. With this magic in place, all plots created in code cells will automatically be displayed inline.

Using the previously created pandas DataFrame that we named df, let's plot the number of countries in each continent as a bar chart using the plot() method of pandas DataFrames.
Now, for the difference between %%sql and %sql: ordinary assignment works for single-line %sql queries while %%sql is for multi-line queries. See the Assignment ipython-sql documentation section for more info.